# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Polideportivo API",
        "description": "Polideportivo API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Get list of reservas",
                "description": "Get a paginated list of reservas or filter by date",
                "operationId": "f64e3c3e32fb3ceb4ef045bfda7f90bc",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "description": "Filter by date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Create a new reserva",
                "description": "Create a new reserva and return it as a JSON response",
                "operationId": "441c2d801167bb64c27408d6f5228906",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-22"
                                    },
                                    "hora": {
                                        "type": "hour",
                                        "example": "11:00"
                                    },
                                    "socio_id": {
                                        "type": "integer",
                                        "example": "22"
                                    },
                                    "pista_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva created successfully"
                    }
                }
            }
        },
        "/api/v1/reservas/{reserva}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Get a specific reserva",
                "description": "Returns a specific reserva as a JSON response",
                "operationId": "1b004cad6adbb9ccd2edec464137f0df",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "description": "ID of the reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successful"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Update a specific reserva",
                "description": "Update a specific reserva and return it as a JSON response",
                "operationId": "74295a99219de6b5ba2ea326597269b5",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "description": "ID of the reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-22"
                                    },
                                    "hora": {
                                        "type": "hour",
                                        "example": "11:00"
                                    },
                                    "socio_id": {
                                        "type": "integer",
                                        "example": "22"
                                    },
                                    "pista_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Delete a specific reserva",
                "description": "Delete a specific reserva and return a JSON response",
                "operationId": "0a0c312556f59fcddd0dcce60e392772",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "description": "ID of the reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva deleted successfully"
                    },
                    "404": {
                        "description": "Reserva not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "0": []
        }
    }
}